+++ src/main.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import "@radix-ui/themes/styles.css";
 import App from "./App";
 import "./styles.css";
 import "./responsive-mobile-tablet.css";